+++ src/main/java/io/games/play_book/game_result/repos/GameResultRepository.java
@@ -6,7 +6,7 @@ import io.games.play_book.game.domain.Game;
 import io.games.play_book.game_result.domain.GameResult;
 import io.games.play_book.season_player.domain.SeasonPlayer;
 
-public interface GameResultRepository extends JpaRepository<GameResult, Integer> {
+public interface GameResultRepository extends JpaRepository<GameResult, Long> {
 
   GameResult findFirstByGame(Game game);
