+++ src/main/java/io/games/play_book/player_participation/repos/PlayerParticipationRepository.java
@@ -6,7 +6,7 @@ import io.games.play_book.game.domain.Game;
 import io.games.play_book.player_participation.domain.PlayerParticipation;
 import io.games.play_book.season_player.domain.SeasonPlayer;
 
-public interface PlayerParticipationRepository extends JpaRepository<PlayerParticipation, Integer> {
+public interface PlayerParticipationRepository extends JpaRepository<PlayerParticipation, Long> {
 
   PlayerParticipation findFirstByGame(Game game);
